refactor: move snapshot to recorder
