--- packages/snapshot/src/snapshot.ts
@@ -1,42 +1,0 @@
-import { createElement } from '@timecat/virtual-dom'
-import { InfoData, SnapshotRecord, VNode, RecordType } from '@timecat/share'
-import { nodeStore, getRadix64TimeStr } from '@timecat/utils'
-
-function getInitInfo(context: Window): InfoData {
-    const { name, publicId, systemId } = context.document.doctype || {}
-    const doctype = () => ({ name, publicId, systemId } as DocumentType)
-    const href = () => context.location.href
-    const width = () => context.innerWidth
-    const height = () => context.innerHeight
-    const scrollTop = () => context.pageYOffset
-    const scrollLeft = () => context.pageXOffset
-
-    const getFrameElement = () => context.frameElement
-    const frameElement = getFrameElement()
-    const frameId = nodeStore.getNodeId(frameElement) || null
-
-    return {
-        doctype: doctype(),
-        href: href(),
-        scrollTop: scrollTop(),
-        scrollLeft: scrollLeft(),
-        width: width(),
-        height: height(),
-        frameId
-    }
-}
-
-function DOMSnapshot(context: Window): SnapshotRecord {
-    return {
-        type: RecordType.SNAPSHOT,
-        data: {
-            vNode: createElement(context.document.documentElement) as VNode,
-            ...getInitInfo(context)
-        },
-        time: getRadix64TimeStr()
-    }
-}
-
-export const snapshots = {
-    DOMSnapshot
-}
